requirements.txt
albumentations
numpy
opencv-python
tqdm
tensorflow_gpu>=2.3.1
